bonus_util no longer imports health_util
bonus_util will no longer import health_util automatically because it was causing some issues with a "ghost health bar" appearing in levels that don't use the health bar. This might be happening due to the "in Level" variables created by health_util not being removed from the cache when you exit the level.